Monter l outillage au niveau du kit, et regler ce que le niveau 8 revele
Le paquet promettait PHP 8.4 quand la suite annonce 8.3 : un projet en
8.3 installait le kit et pas analytics. Verifie avant de descendre, comme
demande : aucune syntaxe propre a 8.4 dans les 153 fichiers de src, et
aucune dependance directe n exige mieux que 8.3. Laravel 13 lui-meme
demande ^8.3 et accepte Symfony 7.4 ou 8 ; le verrou actuel porte Symfony
8 parce que cette machine tourne en 8.5, pas parce que le code l exige.

PHPUnit 11 monte a 12. paratest bloquait, et sa contrainte est epinglee a
~7.20.0 : c est la seule version qui accepte a la fois PHPUnit 12 et PHP
8.3, la 7.21 abandonnant 8.3. Un ^7.20 laisserait Composer prendre la
7.21 sur une machine recente et casser la promesse sans un mot.

PHPStan passe de 7 a 8, avec les regles strictes et sans ligne de base.
Le includes ecrit a la main disparait : extension-installer monte les
extensions seul, et un includes fait doublon des qu il s en ajoute une.

Une regle stricte est coupee, nommement et avec sa mesure :
dynamicCallOnStaticMethod comptait 153 signalements, tous des appels au
constructeur de requetes d Eloquent, dont aucun n est un defaut. Le kit ne
les rencontre pas, faute de modeles.

Restaient 88 signalements reels, tous traites. Trois meritent d etre
signales pour ce qu ils ont change :

- deux chemins de FunnelStep repondaient differemment, le cas simple
  filtrant sur null et le cas multiple laissant aussi tomber la chaine
  vide et le nom « 0 » ;
- le tableau de bord marketing tombait entierement si une annonce
  disparaissait entre l agregation et sa lecture ; il degrade maintenant
  la ligne, comme la ligne des campagnes juste au-dessus le faisait deja ;
- l anonymisation d IP garde une lecture permissive la ou le commutateur
  general passe en === true : pour un reglage de vie privee, une valeur
  ecrite 1 ne doit pas faire stocker l IP brute.

+++ src/Console/GeoipDownloadCommand.php
@@ -173,7 +173,7 @@ final class GeoipDownloadCommand extends Command
             $remaining = $size;
 
             while ($remaining > 0) {
-                $chunk = fread($in, (int) min(1 << 20, $remaining));
+                $chunk = fread($in, min(1 << 20, $remaining));
 
                 if ($chunk === false || $chunk === '') {
                     throw new RuntimeException('the archive ended before the database did');
